refactor: consolidate upload route under user blueprint (#37)
- Moved upload_bp logic into routes/user.py
- Removed upload_bp registration from __init__.py
- Updated url_for() calls to use user.file_upload
- Verified /file_upload route renders correctly
