--- app/static/uploads/test_data.csv
@@ -1,5 +1,0 @@
-Date,Transaction Type,Coin,Amount,Price (USD)
-2024-04-01,Buy,BTC,0.005,30000
-2024-04-02,Buy,ETH,0.1,2000
-2024-04-03,Sell,BTC,0.0025,31000
-2024-04-04,Buy,ADA,500,0.45
