+++ app/templates/base.html
@@ -23,7 +23,7 @@
         <div class="d-flex gap-4 align-items-center">
           <a href="{{ url_for('index') }}" class="nav-link nav-hover">Home</a>
           {% if current_user.is_authenticated %}
-            <a href="{{ url_for('upload.file_upload') }}" class="nav-link nav-hover">File Upload</a>
+            <a href="{{ url_for('user.file_upload') }}" class="nav-link nav-hover">File Upload</a>
             <a href="{{ url_for('visual') }}" class="nav-link nav-hover">Tax Calculation</a>
             <a href="{{ url_for('share') }}" class="nav-link nav-hover">Share Data</a>
             <a href="{{ url_for('user.logout') }}" class="btn btn-outline-warning text-white fw-medium login-btn">Logout</a>
