+++ app/templates/index.html
@@ -29,7 +29,7 @@
     <div class="container">
     <h1>Crypto Tax, Made Tolerable</h1>
     <p class="tagline">Easily calculate your capital gains tax on crypto — fast, simple, accurate.</p>
-    <a href="{{ url_for('upload.file_upload') }}" class="cta-button">Start Your Estimate</a>
+    <a href="{{ url_for('user.file_upload') }}" class="cta-button">Start Your Estimate</a>
 
     <!-- kept these for now in case someone wants to re-use them on a different page
     <div class="features">
