+++ app/templates/user/file_upload.html
@@ -11,7 +11,7 @@
     </div>
     <div class="row px-4 mt-3 mb-3 flex-grow-1 d-flex">
 
-        <!-- JORDY: KEEPING THIS FOR NOW AS A REFERENCE FOR CSS -->
+        <!--  KEEPING THIS FOR NOW AS A REFERENCE FOR CSS -->
         <!--
         <div class="col upload-area flex-column" id="uploadfile">
             <div class="upload-text">Drag and drop file here</div>
